+++ docs/changelog.md
@@ -20,6 +20,7 @@ The versioning system is basically the following:
 - Warn on unnecessary usage of `op`.
 - Warn on some overly complex expressions involving `bool`s.
 - Warn on `case` expressions with only one arm.
+- Warn on unnecessary usage of `;`.
 
 ## v0.3.14